Policies and middlewares finished

+++ app/Http/Controllers/PermissionController.php
@@ -6,18 +6,24 @@ use App\Role;
 use App\Permission;
 use App\Http\Requests\Permission\StoreRequest;
 use App\Http\Requests\Permission\UpdateRequest;
-use Illuminate\Http\Request;
 
 class PermissionController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('role:' . config('app.admin_role'));
+    }
+
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index()
     {
-        return view('theme.backend.pages.permission.index', [
+        $this->authorize('index',Role::class);
+        return view('theme.backend.pages.permission.index',[
             'permissions' => Permission::all()
         ]);
     }
@@ -26,9 +32,11 @@ class PermissionController extends Controller
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
+        $this->authorize('create',Role::class);
         return view('theme.backend.pages.permission.create',[
             'roles' => Role::all()
         ]);
@@ -37,7 +45,8 @@ class PermissionController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param StoreRequest $request
+     * @param Permission $permission
      * @return \Illuminate\Http\Response
      */
     public function store(StoreRequest $request, Permission $permission)
@@ -49,25 +58,29 @@ class PermissionController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Permission  $permission
+     * @param \App\Permission $permission
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show(Permission $permission)
     {
-        return view('theme.backend.pages.permission.show', [
-            'permission' => $permission,
+        $this->authorize('view', $permission);
+        return view('theme.backend.pages.permission.show',[
+            'permission' => $permission
         ]);
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Permission  $permission
+     * @param \App\Permission $permission
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Permission $permission)
     {
-        return view('theme.backend.pages.permission.edit', [
+        $this->authorize('update', $permission);
+        return view('theme.backend.pages.permission.edit',[
             'permission' => $permission,
             'roles' => Role::all()
         ]);
@@ -76,28 +89,29 @@ class PermissionController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Permission  $permission
+     * @param UpdateRequest $request
+     * @param \App\Permission $permission
      * @return \Illuminate\Http\Response
      */
     public function update(UpdateRequest $request, Permission $permission)
     {
-        $permission = $permission->my_update($request);
-        return redirect()->route('backend.permission.show', $permission);
+        $permission->my_update($request);
+        return redirect()->route('backend.permission.show',$permission);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Permission  $permission
+     * @param \App\Permission $permission
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(Permission $permission)
     {
+        $this->authorize('delete', $permission);
         $role = $permission->role;
         $permission->delete();
-        alert('Éxito', 'El permiso se ha eliminado', 'success')->showConfirmButton();
+        alert('Éxito', 'Permiso eliminado', 'success')->showConfirmButton();
         return redirect()->route('backend.role.show', $role);
-
     }
 }
